Arquivos salvos pelo crawler sao colocados na pasta output. Logica de entrada transferida para o arquivo main.go
